+++ designer/client/src/components/CommentInput.tsx
@@ -1,4 +1,4 @@
-import React, { ChangeEventHandler, DetailedHTMLProps, ReactNode, TextareaHTMLAttributes } from "react";
+import React, { DetailedHTMLProps, TextareaHTMLAttributes } from "react";
 import { useTranslation } from "react-i18next";
 import { TextAreaWithFocus } from "./withFocus";
