+++ designer/client/src/components/graph/fragmentGraph.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import { connect } from "react-redux";
 import { compose } from "redux";
 import * as LayoutUtils from "../../reducers/layoutUtils";
