+++ designer/client/src/components/toolbarComponents/toolbarButtons/index.ts
@@ -1,0 +1,3 @@
+export * from "./ToolbarButton";
+export * from "./ToolbarButtons";
+export * from "./types";
