+++ designer/client/src/components/toolbarSettings/defaultToolbarsConfig.ts
@@ -1,7 +1,7 @@
 /* eslint-disable i18next/no-literal-string */
 import { ToolbarsSide } from "../../reducers/toolbars";
 import { WithId } from "../../types/common";
-import { ButtonsVariant } from "../toolbarComponents/ToolbarButtons";
+import { ButtonsVariant } from "../toolbarComponents/toolbarButtons/index";
 import { BuiltinButtonTypes } from "./buttons";
 import { DEV_TOOLBARS } from "./DEV_TOOLBARS";
 import { ToolbarsConfig } from "./types";
