+++ designer/client/src/components/toolbars/AttachmentsPanel.tsx
@@ -1,4 +1,4 @@
-import { ToolbarWrapper } from "../toolbarComponents/ToolbarWrapper";
+import { ToolbarWrapper } from "../toolbarComponents/toolbarWrapper/ToolbarWrapper";
 import ProcessAttachments from "../ProcessAttachments";
 import React from "react";
 import { useTranslation } from "react-i18next";
