+++ designer/client/src/components/toolbars/VersionsPanel.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from "react-i18next";
 import { useSelector } from "react-redux";
 import { getCapabilities } from "../../reducers/selectors/other";
 import ProcessHistory from "../ProcessHistory";
-import { ToolbarWrapper } from "../toolbarComponents/ToolbarWrapper";
+import { ToolbarWrapper } from "../toolbarComponents/toolbarWrapper/ToolbarWrapper";
 
 export function VersionsPanel(): JSX.Element {
     const { t } = useTranslation();
