+++ designer/client/src/components/toolbars/view/buttons/ZoomInButton.tsx
@@ -1,6 +1,6 @@
 import { useTranslation } from "react-i18next";
 import { useDispatch } from "react-redux";
-import ToolbarButton from "../../../toolbarComponents/ToolbarButton";
+import { ToolbarButton } from "../../../toolbarComponents/toolbarButtons/index";
 import { zoomIn } from "../../../../actions/nk/zoom";
 import React from "react";
 import { useGraph } from "../../../graph/GraphContext";
