+++ designer/client/src/components/toolbars/view/buttons/ZoomOutButton.tsx
@@ -1,6 +1,6 @@
 import { useTranslation } from "react-i18next";
 import { useDispatch } from "react-redux";
-import ToolbarButton from "../../../toolbarComponents/ToolbarButton";
+import { ToolbarButton } from "../../../toolbarComponents/toolbarButtons/index";
 import { zoomOut } from "../../../../actions/nk/zoom";
 import React from "react";
 import { useGraph } from "../../../graph/GraphContext";
